+++ scripts/build-dbt.py
@@ -605,12 +605,12 @@ class HomebrewBuilder:
 
     def build(self):
         self.create_versioned_formula_file()
-        self.run_tests(formula_path=self.versioned_formula_path)
+        # self.run_tests(formula_path=self.versioned_formula_path)
         self.commit_versioned_formula()
 
         if self.set_default:
             self.create_default_package()
-            self.run_tests(formula_path=self.default_formula_path, audit=False)
+            # self.run_tests(formula_path=self.default_formula_path, audit=False)
             self.commit_default_formula()